Guage with non-pair

+++ contracts/interfaces/IBribe.sol
@@ -1,9 +1,15 @@
 pragma solidity 0.8.13;
 
 interface IBribe {
-    function _deposit(uint amount, uint tokenId) external;
-    function _withdraw(uint amount, uint tokenId) external;
-    function getRewardForOwner(uint tokenId, address[] memory tokens) external;
-    function notifyRewardAmount(address token, uint amount) external;
-    function left(address token) external view returns (uint);
+    function _deposit(uint256 amount, uint256 tokenId) external;
+
+    function _withdraw(uint256 amount, uint256 tokenId) external;
+
+    function getRewardForOwner(uint256 tokenId, address[] memory tokens)
+        external;
+
+    function notifyRewardAmount(address token, uint256 amount) external;
+
+    function left(address token) external view returns (uint256);
+
 }
